refactor: establish unidirectional dependency - PIT-663
