+++ package.json
@@ -20,19 +20,15 @@
     "@mui/x-date-pickers": "^8.11.3",
     "@mui/x-date-pickers-pro": "^8.11.3",
     "@tanstack/react-query": "^5.89.0",
-    "@turf/circle": "^7.2.0",
-    "@turf/helpers": "^7.2.0",
     "axios": "^1.12.2",
     "classnames": "^2.5.1",
     "clsx": "^2.1.1",
     "date-fns": "^4.1.0",
-    "dayjs": "^1.11.18",
     "dotenv": "^17.2.2",
     "mapbox-gl": "3.15.0",
     "mui": "^0.0.1",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
-    "react-map-gl": "^8.0.4",
     "react-markdown": "^10.1.0",
     "react-router-dom": "^7.9.1",
     "react-toastify": "^11.0.5",
@@ -46,7 +42,6 @@
     "@types/react-dom": "^19.1.9",
     "@vitejs/plugin-react": "^5.0.0",
     "autoprefixer": "^10.4.21",
-    "globals": "^16.3.0",
     "msw": "^2.11.3",
     "postcss": "^8.5.6",
     "tailwindcss": "^3.4.17",
