--- src/features/mypage/types.ts
@@ -1,21 +1,0 @@
-export interface DistrictInfo {
-  id: number | string;
-  name: string;
-  locked: boolean;
-  description?: string;
-  lat?: number;
-  lng?: number;
-}
-
-export interface CityData {
-  cityName: string;
-  totalDistricts: number;
-  lockedDistricts: number;
-  unlockedDistricts: number;
-  districts: DistrictInfo[];
-}
-
-export interface DistrictLockData {
-  totalKeys: number;
-  cities: CityData[];
-}
